fixup! make start_pos an int

+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/MaliciousEntity.java
@@ -16,6 +16,6 @@ public final class MaliciousEntity {
 	String type;
 	String value;
 	boolean redacted;
-	float startPos;
+	int startPos;
 	Map<String, Object> raw;
 }
